requirements.txt
Django==1.11.7
Pillow==4.3.0
boto3==1.4.8
botocore==1.8.9
dj-database-url==0.4.2
django-multiselectfield==0.1.8
django-storages==1.6.5
django-tinymce==2.6.0
docutils==0.14
jmespath==0.9.3
olefile==0.44
psycopg2==2.7.3.2
python-dateutil==2.6.1
pytz==2017.3
s3transfer==0.1.12
